slight refactor

+++ src/main/java/com/greenfox/jasper/domain/Resource.java
@@ -69,7 +69,7 @@ public class Resource implements Serializable {
         this.type = type;
     }
 
-    public double getAmount() {
+    public float getAmount() {
         return amount;
     }
