+++ src/main/java/com/greenfox/jasper/services/ResourceServices.java
@@ -28,7 +28,7 @@ public class ResourceServices {
     TroopServices troopServices;
 
     // TODO refactor! + consider adding a production field into kingdom/resource entity to make this method execution faster
-    public void calculateResource(int kingdomId) {
+    public void calculateResource(long kingdomId) {
 
         Kingdom kingdom = kingdomServices.findOneById(kingdomId);
 
@@ -94,7 +94,7 @@ public class ResourceServices {
         resourceRepo.save(resource);
     }
 
-    public Resource findOneResource(int resourceId) {
+    public Resource findOneResource(long resourceId) {
         return resourceRepo.findOneById(resourceId);
     }
 
@@ -106,11 +106,11 @@ public class ResourceServices {
         return resourceRepo.findOneByKingdomAndType(kingdomServices.findOneById(kingdomId), type);
     }
 
-    public List<Resource> findAllResourcesByKingdomId(int kingdomId) {
-        return resourceRepo.findAllByKingdom(kingdomServices.findOneById((long) kingdomId));
+    public List<Resource> findAllResourcesByKingdomId(long kingdomId) {
+        return resourceRepo.findAllByKingdom(kingdomServices.findOneById(kingdomId));
     }
 
-    public List<Building> findAllBuildingByKingdomIdAndByType(int kingdomId, String mine) {
+    public List<Building> findAllBuildingByKingdomIdAndByType(long kingdomId, String mine) {
         return buildingServices.findAllBuildingByKingdomIdAndByType(kingdomId, mine);
     }
